refactor: moved components out of /islands
