+++ components/UI/ModalExtendedImage.tsx
