+++ components/food-order/FoodItem.tsx
@@ -1,7 +1,7 @@
 //? Import meal icon to mark how many people a meal can feed
 import MealIcon from "../../assets/MealIcon.tsx";
 //? Import the StyledButton to enable users to add the current FoodItem to the cart
-import StyledButton from "../../islands/StyledButton.tsx";
+import StyledButton from "../UI/StyledButton.tsx";
 //? Import Food type to typecast the data received
 import type { Food } from "../../types/Food.ts";
