+++ components/toys/SyntaxHighlighForm.tsx
@@ -1,5 +1,5 @@
 //? Import button to submit the POST request to the same URL
-import StyledButton from "../../islands/StyledButton.tsx";
+import StyledButton from "../UI/StyledButton.tsx";
 
 //? Export mini form to save highlighted texts
 export default function SyntaxHighlighForm() {
